house-cleaning for UploadAvatar
