+++ frontend/src/views/ComingSoonView.vue
@@ -1,9 +1,7 @@
 <template>
   <p class="message">Coming up soon, be patient please :)</p>
 </template>
-<script setup lang="ts">
-import HomeMenu from "@/components/HomeMenu.vue";
-</script>
+<script setup lang="ts"></script>
 <style scoped>
 .message {
   color: darkcyan;
