+++ note/aileen.txt
@@ -141,7 +141,8 @@ Tutorial:
 ******************************************
 (Fri) 2022-06-03
 	- run.sh: add rebuild database volume
-	
+	- house-cleaning for UploadAvatar
+
 
 
 **************************
